+++ src/rootSaga.js
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { moviesSaga } from "./features/moviesSaga";
+import { moviesSaga } from "./features/popularMovies/moviesSaga";
 import { peopleSaga } from "./features/popularPeople/peopleSaga";
 import { personSaga } from "./features/personDetails/personSaga";
 import { movieDetailsSaga } from "./features/movieDetailsSaga";
